:sparkles: feat: review 기능 완성

+++ store-service/src/main/java/com/example/cloudfour/storeservice/domain/review/controller/ReviewCommonRequestDTO.java
@@ -9,7 +9,7 @@ import java.util.UUID;
 @Builder
 public class ReviewCommonRequestDTO {
     UUID storeId;
-    Integer score;
+    Float score;
     String content;
     String pictureUrl;
 }
